Bench / simulate: new orchestration system
It was fun to play with database events, but it can interfere with the
main work, because starting transactions early is a bad idea. Same with
the detection that all regular workers are fished in the continuous one.

This new system uses a multiprocessing.Queue to pass messages from
the regular workers, and an orchestrator process that listen to them and
fires various global events.

According to the profiler, regular workers waiting for each other to
finish accounts for about 40% of the total time on my dev workstation at this
point (it was around 75% with the previous orchestration). Surely that also
means we should have less variability among work to be done accross regular
workers.