0.8: Bulk of adaptation (unit tests only)
Some of the fixes are shortcuts, like Goods = self.Wms.PhysObj
We get a few Deprecation Warnings
But the tests pass. Now we need to try it for good
and to improve test coverage.

As a side remark, the query to find products that need to be
ordered has been improved, and rather than using the new
grouped quantity queries, we played with a way to bail faster
from the inner query. This still has to be benched against
the grouped quantity query.